+++ tests/tinytest/test_funs_other.R
@@ -383,3 +383,45 @@ expect_error(
     pull(foo),
   "must be an integer"
 )
+
+# na_if
+
+test <- polars::pl$DataFrame(
+  x = c(3, 1, 2, 2, 3),
+  grp = c("A", "A", "A", "B", "")
+)
+
+expect_equal(
+  test |>
+    mutate(foo = na_if(x, 3)) |>
+    pull(foo),
+  c(NA, 1, 2, 2, NA)
+)
+
+expect_equal(
+  test |>
+    mutate(foo = na_if(x, 3), .by = grp) |>
+    pull(foo),
+  c(NA, 1, 2, 2, NA)
+)
+
+expect_equal(
+  test |>
+    mutate(foo = na_if(grp, "")) |>
+    pull(foo),
+  c("A", "A", "A", "B", NA)
+)
+
+expect_equal(
+  test |>
+    mutate(foo = na_if(x, c(3, 1, 1, 1, 1))) |>
+    pull(foo),
+  c(NA, NA, 2, 2, 3)
+)
+
+expect_equal(
+  test |>
+    mutate(foo = na_if(x, NA)) |>
+    pull(foo),
+  c(3, 1, 2, 2, 3)
+)