+++ NEWS.md
@@ -3,7 +3,7 @@
 ## New features
 
 * Add support for `all()`, `any()`, `diff()` `dplyr::consecutive_id()`,   
-  `dplyr::n_distinct()`, `length()`, `rev()`, `unique()`.
+  `dplyr::n_distinct()`, `length()`, `rev()`, `stringr::str_dup()`, `unique()`.
 
 ## Bug fixes